Code-Competence-2/Code/requirements.txt
pandas==2.2.2
pyarrow==15.0.2
fastparquet==2024.2.0
firebase-admin==6.5.0
apache-airflow==2.9.0
google-cloud-storage==2.16.0